last modifcation

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mbousouf <mbousouf@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 15:44:59 by mbousouf          #+#    #+#             */
-/*   Updated: 2022/10/23 23:18:09 by mbousouf         ###   ########.fr       */
+/*   Updated: 2022/11/06 12:03:11 by mbousouf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
